Move InteractionDetails into interact package
