--- api/src/main/java/com/wolfyscript/utilities/gui/InteractionDetails.java
@@ -1,10 +1,0 @@
-package com.wolfyscript.utilities.gui;
-
-public interface InteractionDetails {
-
-    boolean isCancelled();
-
-    InteractionResult.ResultType getResultType();
-
-
-}
